+++ src/components/header/header.module.css
@@ -6,30 +6,9 @@
   width: 100%;
 }
 
-.topHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  background-color: var(--clr-accent-2);
-}
-
-.bottomHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  background-color: var(--clr-bg);
-}
-
 .navigationMenu {
   display: flex;
   gap: 20px;
-  padding: 14px 14px;
-}
-
-.bottomNavigationMenu {
-  display: flex;
   padding: 14px 14px;
 }
 
@@ -50,9 +29,3 @@
     text-decoration: underline;
   }
 }
-
-.input {
-  border: 2px solid var(--clr-primary-300);
-  padding: 8px 20px;
-}
-
