+++ src/components/header/top-header/top-header.module.css
@@ -1,0 +1,7 @@
+.topHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  background-color: var(--clr-accent-2);
+}
