+++ src/components/input/input.module.css
@@ -1,5 +1,5 @@
 .wrapper {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
 }
