+++ src/components/svg-builder/enums.ts
@@ -1,8 +1,6 @@
 export enum IconType {
   Login = 'login',
-  Logout = 'logout',
   Registration = 'registration',
-  Basket = 'basket',
   Logo = 'logo',
   Search = 'search',
 }
