+++ src/components/svg-builder/types.ts
@@ -1,0 +1,6 @@
+import type { IconType } from './enums';
+
+export type IconVariant = Record<
+  IconType,
+  React.FC<React.SVGProps<SVGSVGElement>>
+>;
